+++ Chat/Interfaces/LoginAndRegister/ILogin.cs
@@ -2,5 +2,7 @@
 
 public interface ILogin
 {
-    
+    static abstract Task LoginUserAsync();
+    static abstract bool LoginNewUser(string username, string password);
+
 }
