--- Chat/Interfaces/LoginAndRegister/ILoginAndRegister.cs
@@ -1,10 +1,0 @@
-﻿namespace Chat.Interfaces;
-
-public interface ILoginAndRegister
-{
-    static abstract bool LoginNewUser(string username, string password);
-    static abstract bool RegisterNewUser(string currUsername, string currPassword, string currEmail); 
-    static abstract Task LoginUserAsync();
-    static abstract Task<Task> RegisterUserAsync();
-
-}
