+++ Chat/Interfaces/LoginAndRegister/IRegister.cs
@@ -2,5 +2,7 @@
 
 public interface IRegister
 {
-    
+    static abstract Task<Task> RegisterUserAsync();
+    static abstract bool RegisterNewUser(string currUsername, string currPassword, string currEmail);
+
 }
